extract text page logic to hook
